Changed datatype in Interface and TrackData. Finished integrationtest step 3

+++ ATM/ATM/Interfaces/ITrackData.cs
@@ -12,8 +12,8 @@ namespace ATM.Interfaces
          int X { get; set; }
          int Y { get; set; }
          int Altitude { get; set; }
-         double Velocity { get; set; }
-         double Course { get; set; }
+         int Velocity { get; set; }
+         int Course { get; set; }
 
          DateTime TimeStamp { get; set; }
         string ToString();
